Feat: add conversion from EPSG:4978 to EPSG:4326
